test: restructures tests folder
Adds folder for test_delete_functions and file is changed to
test_ec2_service.py within. Now, each service can have its own test
file.

Additionally, new test are added for delete_ec2_instance to test
1. if instance is in state shutting-down
2. for when its called from delete_autoscaling_group
